+++ Project21_Volume-File/Convert.h
@@ -5,5 +5,5 @@
 using namespace std;
 
 int hexToDecimal(string hex);
-string* decimalToHex(int numByte, int32_t decimal);
+string* decimalToHex(int numByte, uint64_t decimal);
 int charToInt(const char* listChar, int numByte);
